[#114] Fix typescript error from violating next layout file type expectations
I'm not entirely sure what combination of running `next build` and `tsc --noemit` causes this, and why it doesn't seem to repro without some git action (rebase?) in between.

However, I think there's something going on that's like
- There is potentially a state where `.next/` files have not been generated, or not fully generated
- When `npm run build` (`next build`) it is run, it generates the type file `.next/types/app/form/(formSteps)/layout.ts`
- Some git action that may or may not be necessary happens
- Running `npm run typecheck` (`tsc --noemit`) finds the generated `.next/types/app/form/(formSteps)/layout.ts` which specifies a defined set of properties that are expected/allowed on a NextJS Layout page. However, the actual `app/form/(formSteps)/layout.tsx` layout file also exports an additional `FormLayout` which was broken out and exported for testing purposes. tsc errors that `FormLayout` is not a valid thing to be exported

In theory, it seems like NextJS should have thrown an [invalid-segment-export](https://nextjs.org/docs/messages/invalid-segment-export) error. But we haven't seen that; not sure where we should have seen it.

This commit hopefully resolves #[114](https://github.com/newjersey/doula-pm/issues/114).
