+++ src/app/form/(formSteps)/layout.tsx
@@ -1,9 +1,8 @@
 import type { Metadata } from "next";
 import { headers } from "next/headers";
 import React from "react";
-import { allSections, getCurrentStep } from "../_utils/sections";
-
-type CompletionState = "complete" | "current" | "incomplete";
+import { getCurrentStep } from "../_utils/sections";
+import { FormLayout } from "./FormLayout";
 
 export const generateMetadata = async (): Promise<Metadata> => {
   const headersList = await headers();
@@ -18,82 +17,6 @@ export const generateMetadata = async (): Promise<Metadata> => {
   };
 };
 
-export const FormLayout = (props: { children?: React.ReactNode; pathname: string }) => {
-  const { section: currentSection, stepNum: currentStepNum } = getCurrentStep(props.pathname);
-  const currentSectionIndex = allSections.findIndex(
-    (sections) => sections.id === currentSection.id,
-  );
-
-  return (
-    <>
-      <div className="usa-step-indicator" aria-label="progress">
-        <ol className="usa-step-indicator__segments">
-          {allSections.map((sections, sectionIndex) => {
-            let completionState: CompletionState;
-            switch (true) {
-              case sectionIndex < currentSectionIndex:
-                completionState = "complete";
-                break;
-              case sectionIndex === currentSectionIndex:
-                completionState = "current";
-                break;
-              case sectionIndex > currentSectionIndex:
-                completionState = "incomplete";
-                break;
-              default:
-                throw new Error(`Unexpected logic path: ${sectionIndex}, ${currentSectionIndex}`);
-                break;
-            }
-
-            const liSegmentClassSuffix = {
-              complete: "complete",
-              current: "current",
-              incomplete: null,
-            }[completionState];
-            const screenreaderStatus = {
-              complete: "completed",
-              current: null,
-              incomplete: "not completed",
-            }[completionState];
-
-            return (
-              <li
-                key={sections.id}
-                className={`usa-step-indicator__segment ${liSegmentClassSuffix ? `usa-step-indicator__segment--${liSegmentClassSuffix}` : ""}`}
-                {...(completionState === "current" && { "aria-current": "true" })}
-              >
-                <span className="usa-step-indicator__segment-label">
-                  {sections.sectionName}
-                  {screenreaderStatus && <span className="usa-sr-only">{screenreaderStatus}</span>}
-                </span>
-              </li>
-            );
-          })}
-        </ol>
-
-        <div className="usa-step-indicator__header">
-          <h1 className="font-heading-lg">
-            {currentStepNum !== null && (
-              <span className="usa-step-indicator__heading-counter">
-                <span className="usa-sr-only" data-testid="step-text">
-                  Step
-                </span>
-                <span className="usa-step-indicator__current-step">{currentStepNum}</span>
-                &nbsp;
-                <span className="usa-step-indicator__total-steps">{`of ${currentSection.numSteps}`}</span>
-                &nbsp;
-              </span>
-            )}
-            <span className="usa-step-indicator__heading-text">{currentSection.heading}</span>
-          </h1>
-        </div>
-      </div>
-      <div className="margin-top-4">{props.children}</div>
-    </>
-  );
-};
-
-// Separated this into a wrapper because as of writing, Jest does not support testing NextJs asynchronous server components (https://nextjs.org/docs/app/guides/testing/jest)
 const FormLayoutWithRequestContext = async ({ children }: { children?: React.ReactNode }) => {
   const headersList = await headers();
   const pathname = headersList.get("x-pathname") as string;
